Split gawk compatibility suites by test family
